requirements.txt
pandas
google-cloud-storage
functions-framework==3.*
pyarrow
google-cloud-bigquery==3.10.0